Don't consider extended records with flash address 0
Usually the hex files generated from code written in assembly,
or exported after read from a device, have the first line as
an extended record with the address 0. Normally, when we detect an
extended record we ignore all data after it, until the end.
This modification avoids setting the "extended_rec_started" variable if
the flash address is 0, because this means that it is the first hex line,
and we don't want to ignore all lines of the hex.