Fix for evaluating extended records
Fixing implementation from commit 634a6d90b7233abb23aeb250e3c1c9829b4ac88b.
We don't want to set the extended record flag when the address represented in the
two data bytes of the extended record is 0000.
Before this fix, instead of the two data bytes of the record
we were evaluating the address of the record which was always 0000.