Fix flashing for devices with flash size >32KB
Fixed bug where we decide based upon the negative status flag if
the address to be flashed is inside the bootloader space.
sublw sets negative flag if W > k literal(negative result) but also
if (W < k) AND (k - w) > 0x7f.
For a device with flash size 64KB, "btld_code_start_hi - 1" is 0xfc
and subtracting the address 0 will result in 0xfc that will set the
negative flag.
We instead use the CARRY flag interpreted as /BORROW in order to
decide if we subtract an address higher or lower than the bootloader
start address

https://github.com/spanceac/PIC18-bootloader/issues/1